Initial refactoring of speech rendering into dedicated audio module.
