+++ src/audio/acss_renderer.js
@@ -1,0 +1,148 @@
+// Copyright 2017 Volker Sorge
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+/**
+ * @fileoverview An abstract class for audio renderer with prosody markup.
+ *
+ * @author volker.sorge@gmail.com (Volker Sorge)
+ */
+goog.provide('sre.AcssRenderer');
+
+goog.require('sre.AudioUtil');
+goog.require('sre.MarkupRenderer');
+
+
+// /**
+//  * Translates a list of auditory descriptions into a string with SSML markup.
+//  * Currently returns an sexp for emacs speak.
+//  * @param {!Array.<sre.AuditoryDescription>} descrs The list of descriptions.
+//  * @param {string} separator The separator string.
+//  * @return {string} The generated string with ACSS markup.
+//  */
+
+/**
+ * @constructor
+ * @extends {sre.MarkupRenderer}
+ */
+sre.AcssRenderer = function() {
+  sre.AcssRenderer.base(this, 'constructor');
+};
+goog.inherits(sre.AcssRenderer, sre.MarkupRenderer);
+
+
+/**
+ * @override
+ */
+sre.AcssRenderer.prototype.markup = function(descrs) {
+  this.setScaleFunction(-2, 2, 0, 10, 0);
+  var markup = sre.AudioUtil.personalityMarkup(descrs);
+  var result = [];
+  var currentPers = {open: []};
+  var pause = null;
+  var string = false;
+  for (var i = 0, descr; descr = markup[i]; i++) {
+    if (sre.AudioUtil.isMarkupElement(descr)) {
+      sre.AudioUtil.mergeMarkup(currentPers, descr);
+      continue;
+    }
+    if (sre.AudioUtil.isPauseElement(descr)) {
+      if (string) {
+        pause = sre.AudioUtil.mergePause(
+            pause,
+            /** @type {{pause: number}} */(descr), Math.max);
+      }
+      continue;
+    }
+    var str = '"' + descr.string + '"';
+    string = true;
+    if (pause) {
+      result.push(this.sexpPause_(pause));
+      pause = null;
+    }
+    var prosody = this.sexpProsody_(currentPers);
+    result.push(prosody ? '(text (' + prosody + ') ' + str + ')' : str);
+  }
+  return '(exp ' + result.join(this.getSeparator()) + ')';
+};
+
+
+/**
+ * @override
+ */
+sre.AcssRenderer.prototype.merge = function(strs) {
+  return '(exp ' +
+    strs.map(function(str) {
+      return str.replace(/^\(exp /, '').replace(/\)$/, '');}).join(' ') +
+    ')';
+};
+
+
+/**
+ * @override
+ */
+sre.AcssRenderer.prototype.error = function(key) {
+  return '(error "' + key + '")';
+};
+
+
+// TODO: Refactor into methods on all markup renderers.
+/**
+ * Transforms a prosody element into an S-expression.
+ * @param {Object.<string, number>} pros The prosody element.
+ * @return {string} The S-expression.
+ * @private
+ */
+sre.AcssRenderer.prototype.sexpProsody_ = function(pros) {
+  var keys = pros.open;
+  var result = [];
+  for (var i = 0, key; key = keys[i]; i++) {
+    result.push(this.sexpProsodyElement_(key, pros[key]));
+  }
+  return result.join(' ');
+};
+
+
+/**
+ * Transforms a prosody key value pair into an S-expression.
+ * @param {sre.Engine.personalityProps} key The prosody name.
+ * @param {number} value The prosody value.
+ * @return {string} The S-expression.
+ * @private
+ */
+sre.AcssRenderer.prototype.sexpProsodyElement_ = function(key, value) {
+  value = this.applyScaleFunction(value);
+  switch (key) {
+    case sre.Engine.personalityProps.RATE:
+      return '(richness . ' + value + ')';
+    case sre.Engine.personalityProps.PITCH:
+      return '(average-pitch . ' + value + ')';
+      break;
+    case sre.Engine.personalityProps.VOLUME:
+      return '(stress . ' + value + ')';
+      break;
+  }
+  return '(value . ' + value + ')';
+};
+
+
+/**
+ * Translates a pause into its corresponding S-expression.
+ * @param {{pause: number}} pause A pause element.
+ * @return {string} The S-expression.
+ * @private
+ */
+sre.AcssRenderer.prototype.sexpPause_ = function(pause) {
+  return '(pause . ' + pause[sre.Engine.personalityProps.PAUSE] + ')';
+};
