+++ src/speech_generator/abstract_speech_generator.js
@@ -22,7 +22,7 @@
 
 goog.provide('sre.AbstractSpeechGenerator');
 
-goog.require('sre.AuditoryDescription');
+goog.require('sre.AuralRendering');
 goog.require('sre.RebuildStree');
 goog.require('sre.SpeechGenerator');
 goog.require('sre.SpeechGeneratorUtil');
@@ -88,5 +88,5 @@ sre.AbstractSpeechGenerator.prototype.generateSpeech = function(node, xml) {
     this.rebuilt_ = new sre.RebuildStree(xml);
   }
   var descrs = sre.SpeechGeneratorUtil.computeSpeech(this.getRebuilt().xml);
-  return sre.AuditoryDescription.speechString(descrs);
+  return sre.AuralRendering.getInstance().markup(descrs);
 };
