+++ src/walker/abstract_walker.js
@@ -22,6 +22,7 @@
 goog.provide('sre.AbstractWalker');
 
 goog.require('sre.AuditoryDescription');
+goog.require('sre.AuralRendering');
 goog.require('sre.DomUtil');
 goog.require('sre.EnrichMathml.Attribute');
 goog.require('sre.EventUtil.KeyCode');
@@ -234,7 +235,7 @@ sre.AbstractWalker.prototype.speech = function() {
                 sre.SpeechGeneratorUtil.retrieveSpeech(this.xml, snode));
   }
   if (prefix) speech.unshift(prefix);
-  return sre.AuditoryDescription.mergeStrings(speech);
+  return sre.AuralRendering.getInstance().merge(speech);
 };
 
 
@@ -250,7 +251,7 @@ sre.AbstractWalker.prototype.levelAnnouncement_ = function(prefix) {
   var primary = this.focus_.getDomPrimary();
   var expand = (this.expandable(primary) && 'expandable') ||
       (this.collapsible(primary) && 'collapsible') || '';
-  var descr = [sre.AuditoryDescription.speechString(
+  var descr = [sre.AuralRendering.getInstance().markup(
       [new sre.AuditoryDescription({text: 'Level ' + this.getDepth(),
          personality: {}})])];
   if (prefix) {
@@ -259,7 +260,7 @@ sre.AbstractWalker.prototype.levelAnnouncement_ = function(prefix) {
   if (expand) {
     descr.push(expand);
   }
-  return sre.AuditoryDescription.mergeStrings(descr);
+  return sre.AuralRendering.getInstance().merge(descr);
 };
 
 
@@ -556,5 +557,3 @@ sre.AbstractWalker.prototype.singletonFocus = function(id) {
 sre.AbstractWalker.prototype.focusFromId = function(id, ids) {
   return sre.Focus.factory(id, ids, this.rebuilt, this.node);
 };
-
-
